+++ .julia/project.jl/start.jl
@@ -1,9 +1,0 @@
-import Pkg
-
-packages = [
-  "Revise"
-]
-Pkg.activate(".")
-Pkg.add(packages)
-
-using Revise