+++ .r-hub/validate-email.R
@@ -1,1 +1,0 @@
-rhub::validate_email(email = "r.jeksterslab@gmail.com")