+++ R/semmcci-pcci-dot.R
@@ -9,7 +9,8 @@
 #' @param alpha Numeric vector.
 #'   Significance level.
 #'   Default value is `alpha = c(0.001, 0.01, 0.05)`.
-#' @return Returns a matrix of estimates, standard errors, number of Monte Carlo replications, and confidence intervals.
+#' @return Returns a matrix of estimates, standard errors,
+#'   number of Monte Carlo replications, and confidence intervals.
 #' @keywords internal
 #' @noRd
 .PCCI <- function(thetahatstar,