+++ tests/testthat/test-semmcci-mc-latent-med-defined-none.R
@@ -19,13 +19,6 @@ lapply(
       data = data
     )
     set.seed(seed)
-    results_null <- MC(
-      fit,
-      R = R,
-      alpha = c(0.001, 0.01, 0.05),
-      decomposition = NULL
-    )
-    set.seed(seed)
     results_chol <- MC(
       fit,
       R = R,
@@ -53,21 +46,6 @@ lapply(
       Sigma = lavaan::vcov(fit)
     )
     testthat::test_that(
-      paste(text, "NULL"),
-      {
-        testthat::expect_equal(
-          results_null$thetahat$est[names(lavaan::coef(fit))],
-          as.vector(lavaan::coef(fit)),
-          check.attributes = FALSE
-        )
-        testthat::expect_true(
-          abs(
-            .MCCI(results_null)["visual~~textual", "97.5%"] - quantile(answers[, "visual~~textual"], .975, na.rm = TRUE)
-          ) <= tol
-        )
-      }
-    )
-    testthat::test_that(
       paste(text, "chol"),
       {
         testthat::expect_equal(
@@ -77,7 +55,13 @@ lapply(
         )
         testthat::expect_true(
           abs(
-            .MCCI(results_chol)["visual~~textual", "97.5%"] - quantile(answers[, "visual~~textual"], .975, na.rm = TRUE)
+            .MCCI(
+              results_chol
+            )["visual~~textual", "97.5%"] - quantile(
+              answers[, "visual~~textual"],
+              .975,
+              na.rm = TRUE
+            )
           ) <= tol
         )
       }
@@ -92,7 +76,13 @@ lapply(
         )
         testthat::expect_true(
           abs(
-            .MCCI(results_eigen)["visual~~textual", "97.5%"] - quantile(answers[, "visual~~textual"], .975, na.rm = TRUE)
+            .MCCI(
+              results_eigen
+            )["visual~~textual", "97.5%"] - quantile(
+              answers[, "visual~~textual"],
+              .975,
+              na.rm = TRUE
+            )
           ) <= tol
         )
       }
@@ -107,7 +97,13 @@ lapply(
         )
         testthat::expect_true(
           abs(
-            .MCCI(results_svd)["visual~~textual", "97.5%"] - quantile(answers[, "visual~~textual"], .975, na.rm = TRUE)
+            .MCCI(
+              results_svd
+            )["visual~~textual", "97.5%"] - quantile(
+              answers[, "visual~~textual"],
+              .975,
+              na.rm = TRUE
+            )
           ) <= tol
         )
       }