+++ tests/testthat/test-semmcci-mc-latent-med-defined.R
@@ -21,13 +21,6 @@ lapply(
       data = data
     )
     set.seed(seed)
-    results_null <- MC(
-      fit,
-      R = R,
-      alpha = c(0.001, 0.01, 0.05),
-      decomposition = NULL
-    )
-    set.seed(seed)
     results_chol <- MC(
       fit,
       R = R,
@@ -59,21 +52,6 @@ lapply(
       ab = answers[, "a"] * answers[, "b"]
     )
     testthat::test_that(
-      paste(text, "NULL"),
-      {
-        testthat::expect_equal(
-          results_null$thetahat$est,
-          lavaan::parameterEstimates(fit)$est,
-          check.attributes = FALSE
-        )
-        testthat::expect_true(
-          abs(
-            .MCCI(results_null)["ab", "97.5%"] - quantile(answers[, "ab"], .975, na.rm = TRUE)
-          ) <= tol
-        )
-      }
-    )
-    testthat::test_that(
       paste(text, "chol"),
       {
         testthat::expect_equal(
@@ -83,7 +61,13 @@ lapply(
         )
         testthat::expect_true(
           abs(
-            .MCCI(results_chol)["ab", "97.5%"] - quantile(answers[, "ab"], .975, na.rm = TRUE)
+            .MCCI(
+              results_chol
+            )["ab", "97.5%"] - quantile(
+              answers[, "ab"],
+              .975,
+              na.rm = TRUE
+            )
           ) <= tol
         )
       }
@@ -98,7 +82,13 @@ lapply(
         )
         testthat::expect_true(
           abs(
-            .MCCI(results_eigen)["ab", "97.5%"] - quantile(answers[, "ab"], .975, na.rm = TRUE)
+            .MCCI(
+              results_eigen
+            )["ab", "97.5%"] - quantile(
+              answers[, "ab"],
+              .975,
+              na.rm = TRUE
+            )
           ) <= tol
         )
       }
@@ -113,7 +103,13 @@ lapply(
         )
         testthat::expect_true(
           abs(
-            .MCCI(results_svd)["ab", "97.5%"] - quantile(answers[, "ab"], .975, na.rm = TRUE)
+            .MCCI(
+              results_svd
+            )["ab", "97.5%"] - quantile(
+              answers[, "ab"],
+              .975,
+              na.rm = TRUE
+            )
           ) <= tol
         )
       }