+++ tests/testthat/test-semmcci-mc-latent-med-std-defined-none.R
@@ -20,16 +20,6 @@ lapply(
       data = data
     )
     set.seed(seed)
-    results_unstd_null <- MC(
-      fit,
-      R = R,
-      alpha = c(0.001, 0.01, 0.05),
-      decomposition = NULL
-    )
-    # insert original estimate on the third row
-    results_unstd_null$thetahatstar[3, ] <- lavaan::parameterEstimates(fit)$est
-    results_null <- MCStd(results_unstd_null)
-    set.seed(seed)
     results_unstd_chol <- MC(
       fit,
       R = R,
@@ -60,26 +50,6 @@ lapply(
     results_unstd_svd$thetahatstar[3, ] <- lavaan::parameterEstimates(fit)$est
     results_svd <- MCStd(results_unstd_svd)
     testthat::test_that(
-      paste(text, "NULL"),
-      {
-        testthat::expect_equal(
-          results_null$thetahat$est,
-          lavaan::parameterEstimates(fit)$est,
-          check.attributes = FALSE
-        )
-        testthat::expect_equal(
-          results_null$thetahatstar_std[3, ],
-          lavaan::standardizedSolution(fit)$est.std,
-          check.attributes = FALSE
-        )
-        testthat::expect_equal(
-          .MCCI(results_null)["textual~visual", "97.5%"],
-          quantile(results_null$thetahatstar_std[, "textual~visual"], .975, na.rm = TRUE),
-          check.attributes = FALSE
-        )
-      }
-    )
-    testthat::test_that(
       paste(text, "chol"),
       {
         testthat::expect_equal(
@@ -93,8 +63,14 @@ lapply(
           check.attributes = FALSE
         )
         testthat::expect_equal(
-          .MCCI(results_chol)["textual~visual", "97.5%"],
-          quantile(results_chol$thetahatstar_std[, "textual~visual"], .975, na.rm = TRUE),
+          .MCCI(
+            results_chol
+          )["textual~visual", "97.5%"],
+          quantile(
+            results_chol$thetahatstar_std[, "textual~visual"],
+            .975,
+            na.rm = TRUE
+          ),
           check.attributes = FALSE
         )
       }
@@ -113,8 +89,14 @@ lapply(
           check.attributes = FALSE
         )
         testthat::expect_equal(
-          .MCCI(results_eigen)["textual~visual", "97.5%"],
-          quantile(results_eigen$thetahatstar_std[, "textual~visual"], .975, na.rm = TRUE),
+          .MCCI(
+            results_eigen
+          )["textual~visual", "97.5%"],
+          quantile(
+            results_eigen$thetahatstar_std[, "textual~visual"],
+            .975,
+            na.rm = TRUE
+          ),
           check.attributes = FALSE
         )
       }
@@ -133,8 +115,14 @@ lapply(
           check.attributes = FALSE
         )
         testthat::expect_equal(
-          .MCCI(results_svd)["textual~visual", "97.5%"],
-          quantile(results_svd$thetahatstar_std[, "textual~visual"], .975, na.rm = TRUE),
+          .MCCI(
+            results_svd
+          )["textual~visual", "97.5%"],
+          quantile(
+            results_svd$thetahatstar_std[, "textual~visual"],
+            .975,
+            na.rm = TRUE
+          ),
           check.attributes = FALSE
         )
       }