+++ tests/testthat/test-semmcci-mc-simple-med-defined-none.R
@@ -29,13 +29,6 @@ lapply(
       fixed.x = FALSE
     )
     set.seed(seed)
-    results_null <- MC(
-      fit,
-      R = R,
-      alpha = c(0.001, 0.01, 0.05),
-      decomposition = NULL
-    )
-    set.seed(seed)
     results_chol <- MC(
       fit,
       R = R,
@@ -65,21 +58,6 @@ lapply(
       answers
     )
     testthat::test_that(
-      paste(text, "NULL"),
-      {
-        testthat::expect_equal(
-          results_null$thetahat$est,
-          lavaan::parameterEstimates(fit)$est,
-          check.attributes = FALSE
-        )
-        testthat::expect_true(
-          abs(
-            .MCCI(results_null)["cp", "97.5%"] - quantile(answers[, "cp"], .975, na.rm = TRUE)
-          ) <= tol
-        )
-      }
-    )
-    testthat::test_that(
       paste(text, "chol"),
       {
         testthat::expect_equal(
@@ -89,7 +67,13 @@ lapply(
         )
         testthat::expect_true(
           abs(
-            .MCCI(results_chol)["cp", "97.5%"] - quantile(answers[, "cp"], .975, na.rm = TRUE)
+            .MCCI(
+              results_chol
+            )["cp", "97.5%"] - quantile(
+              answers[, "cp"],
+              .975,
+              na.rm = TRUE
+            )
           ) <= tol
         )
       }
@@ -104,7 +88,12 @@ lapply(
         )
         testthat::expect_true(
           abs(
-            .MCCI(results_eigen)["cp", "97.5%"] - quantile(answers[, "cp"], .975)
+            .MCCI(
+              results_eigen
+            )["cp", "97.5%"] - quantile(
+              answers[, "cp"],
+              .975
+            )
           ) <= tol
         )
       }
@@ -119,7 +108,13 @@ lapply(
         )
         testthat::expect_true(
           abs(
-            .MCCI(results_svd)["cp", "97.5%"] - quantile(answers[, "cp"], .975, na.rm = TRUE)
+            .MCCI(
+              results_svd
+            )["cp", "97.5%"] - quantile(
+              answers[, "cp"],
+              .975,
+              na.rm = TRUE
+            )
           ) <= tol
         )
       }