+++ tests/testthat/test-semmcci-mc-simple-med-std-defined.R
@@ -35,15 +35,6 @@ lapply(
       fixed.x = FALSE
     )
     set.seed(seed)
-    results_unstd_null <- MC(
-      fit,
-      R = R,
-      alpha = c(0.001, 0.01, 0.05),
-      decomposition = NULL
-    )
-    results_unstd_null$thetahatstar[3, ] <- lavaan::parameterEstimates(fit)$est
-    results_null <- MCStd(results_unstd_null)
-    set.seed(seed)
     results_unstd_chol <- MC(
       fit,
       R = R,
@@ -71,26 +62,6 @@ lapply(
     results_unstd_svd$thetahatstar[3, ] <- lavaan::parameterEstimates(fit)$est
     results_svd <- MCStd(results_unstd_svd)
     testthat::test_that(
-      paste(text, "NULL"),
-      {
-        testthat::expect_equal(
-          results_null$thetahat$est,
-          lavaan::parameterEstimates(fit)$est,
-          check.attributes = FALSE
-        )
-        testthat::expect_equal(
-          results_null$thetahatstar_std[3, ],
-          lavaan::standardizedSolution(fit)$est.std,
-          check.attributes = FALSE
-        )
-        testthat::expect_equal(
-          .MCCI(results_null)["ab", "97.5%"],
-          quantile(results_null$thetahatstar_std[, "ab"], .975, na.rm = TRUE),
-          check.attributes = FALSE
-        )
-      }
-    )
-    testthat::test_that(
       paste(text, "chol"),
       {
         testthat::expect_equal(
@@ -104,8 +75,14 @@ lapply(
           check.attributes = FALSE
         )
         testthat::expect_equal(
-          .MCCI(results_chol)["ab", "97.5%"],
-          quantile(results_chol$thetahatstar_std[, "ab"], .975, na.rm = TRUE),
+          .MCCI(
+            results_chol
+          )["ab", "97.5%"],
+          quantile(
+            results_chol$thetahatstar_std[, "ab"],
+            .975,
+            na.rm = TRUE
+          ),
           check.attributes = FALSE
         )
       }
@@ -124,8 +101,14 @@ lapply(
           check.attributes = FALSE
         )
         testthat::expect_equal(
-          .MCCI(results_eigen)["ab", "97.5%"],
-          quantile(results_eigen$thetahatstar_std[, "ab"], .975, na.rm = TRUE),
+          .MCCI(
+            results_eigen
+          )["ab", "97.5%"],
+          quantile(
+            results_eigen$thetahatstar_std[, "ab"],
+            .975,
+            na.rm = TRUE
+          ),
           check.attributes = FALSE
         )
       }
@@ -144,8 +127,14 @@ lapply(
           check.attributes = FALSE
         )
         testthat::expect_equal(
-          .MCCI(results_svd)["ab", "97.5%"],
-          quantile(results_svd$thetahatstar_std[, "ab"], .975, na.rm = TRUE),
+          .MCCI(
+            results_svd
+          )["ab", "97.5%"],
+          quantile(
+            results_svd$thetahatstar_std[, "ab"],
+            .975,
+            na.rm = TRUE
+          ),
           check.attributes = FALSE
         )
       }